# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Clover_TH API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/admin/auctions/rounds": {
      "post": {
        "tags": [
          "auctions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "LIVE_CLAIM",
                      "QUEUE_RANKED"
                    ]
                  },
                  "name": {
                    "type": "string",
                    "maxLength": 400
                  },
                  "durationSec": {
                    "default": 300,
                    "type": "integer",
                    "minimum": 5,
                    "maximum": 86400
                  },
                  "winCap": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 50
                  },
                  "startDelaySec": {
                    "default": 3,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 60
                  },
                  "items": {
                    "maxItems": 500,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "maxLength": 400
                        },
                        "category": {
                          "type": "string",
                          "enum": [
                            "PET",
                            "MATERIAL",
                            "GEMBOX",
                            "GEAR",
                            "CARD",
                            "RELIC"
                          ]
                        },
                        "rarity": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 128
                        },
                        "imageUrl": {
                          "nullable": true,
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 2000
                        }
                      },
                      "required": [
                        "name",
                        "category"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "required": [
                  "type",
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "LIVE_CLAIM",
                        "QUEUE_RANKED"
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "OPEN",
                        "CLOSED",
                        "CANCELLED"
                      ]
                    },
                    "durationSec": {
                      "type": "number"
                    },
                    "winCap": {
                      "nullable": true,
                      "type": "number"
                    },
                    "startDelaySec": {
                      "type": "number"
                    },
                    "opensAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "closesAt": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "type",
                    "name",
                    "status",
                    "durationSec",
                    "winCap",
                    "startDelaySec",
                    "opensAt",
                    "closesAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auctions/rounds/{id}": {
      "patch": {
        "tags": [
          "auctions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 400
                  },
                  "durationSec": {
                    "type": "integer",
                    "minimum": 5,
                    "maximum": 86400
                  },
                  "winCap": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 50
                  },
                  "startDelaySec": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 60
                  },
                  "items": {
                    "maxItems": 500,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "maxLength": 400
                        },
                        "category": {
                          "type": "string",
                          "enum": [
                            "PET",
                            "MATERIAL",
                            "GEMBOX",
                            "GEAR",
                            "CARD",
                            "RELIC"
                          ]
                        },
                        "rarity": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 128
                        },
                        "imageUrl": {
                          "nullable": true,
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 2000
                        }
                      },
                      "required": [
                        "name",
                        "category"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "LIVE_CLAIM",
                        "QUEUE_RANKED"
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "OPEN",
                        "CLOSED",
                        "CANCELLED"
                      ]
                    },
                    "durationSec": {
                      "type": "number"
                    },
                    "winCap": {
                      "nullable": true,
                      "type": "number"
                    },
                    "startDelaySec": {
                      "type": "number"
                    },
                    "opensAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "closesAt": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "type",
                    "name",
                    "status",
                    "durationSec",
                    "winCap",
                    "startDelaySec",
                    "opensAt",
                    "closesAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auctions/rounds/{id}/start": {
      "post": {
        "tags": [
          "auctions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "default": {},
                "type": "object",
                "properties": {
                  "startDelaySec": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 60
                  },
                  "durationSec": {
                    "type": "integer",
                    "minimum": 5,
                    "maximum": 86400
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "LIVE_CLAIM",
                        "QUEUE_RANKED"
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "OPEN",
                        "CLOSED",
                        "CANCELLED"
                      ]
                    },
                    "durationSec": {
                      "type": "number"
                    },
                    "winCap": {
                      "nullable": true,
                      "type": "number"
                    },
                    "startDelaySec": {
                      "type": "number"
                    },
                    "opensAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "closesAt": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "type",
                    "name",
                    "status",
                    "durationSec",
                    "winCap",
                    "startDelaySec",
                    "opensAt",
                    "closesAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auctions/rounds/{id}/close": {
      "post": {
        "tags": [
          "auctions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "LIVE_CLAIM",
                        "QUEUE_RANKED"
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "OPEN",
                        "CLOSED",
                        "CANCELLED"
                      ]
                    },
                    "durationSec": {
                      "type": "number"
                    },
                    "winCap": {
                      "nullable": true,
                      "type": "number"
                    },
                    "startDelaySec": {
                      "type": "number"
                    },
                    "opensAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "closesAt": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "type",
                    "name",
                    "status",
                    "durationSec",
                    "winCap",
                    "startDelaySec",
                    "opensAt",
                    "closesAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/auctions/rounds/{id}/preferences": {
      "get": {
        "tags": [
          "auctions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "roundId": {
                      "type": "number"
                    },
                    "lists": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "memberId": {
                            "type": "string"
                          },
                          "itemIds": {
                            "type": "array",
                            "items": {
                              "type": "number"
                            }
                          }
                        },
                        "required": [
                          "memberId",
                          "itemIds"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "roundId",
                    "lists"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}